Update Canada_Wrangling file, and add a new document that wrangles and one that plots the data
I removed an unneccessary line from the FoodImports_Canada_Wrangling file, and then I created a new data set by wrangling it so that it is only the top 10 countries by most total food imported value in billions of dollars. The last file just creates a graph for the second file.